Porting to macOS.
Without CMAKE_MACOS_BUNDLE no notification is shown, which took some
time to figure out.
All notifications are truncated to a length.
Left and right click are the same. Both will emit the 'activated'
signal as well as show the context menu. This required a slightly
different approach to the UI, in how to get a fortune.